Fix flaky ivy completion
I was getting frustrated that ivy completion was not always available
for find-file and org-roam searches. It turns out that ivy would be
properly activated by use of /counsel/ commands such as C-s and C-x
C-b and work thereafter.

Presumably this is related to how use-package deferred loads work;
replacing =completing-read-function= is a side-effect of ivy
activation and not a registered function itself, so it sits there
latently until something else files off (ivy-mode). This post-init
hook seems to work.