Basic testing with bats
Testing emacs directly within a --batch call is tricky because many
functions are unavailable in this "isinteractive nil" mode. The
workaround is to send lisp to a running emacs daemon instead, which
provides the interactive frame.

This can be done from another emacs with server-eval-at but it's
tricky to get all the error/warning side-effects out. Much simpler to
do it with emacsclient!